feat(local-blockchain): seperate local-blockchain into it's own module
