+++ src/lib/mina.ts
@@ -1,31 +1,13 @@
-import { Ledger, Test } from '../snarky.js';
+import { Test } from '../snarky.js';
 import { Field } from './core.js';
 import { UInt32, UInt64 } from './int.js';
-import { PrivateKey, PublicKey } from './signature.js';
-import {
-  ZkappCommand,
-  AccountUpdate,
-  ZkappPublicInput,
-  TokenId,
-  Authorization,
-  Actions,
-  Events,
-  dummySignature,
-} from './account-update.js';
+import { PublicKey } from './signature.js';
+import { ZkappCommand, TokenId, Authorization } from './account-update.js';
 import * as Fetch from './fetch.js';
-import { NetworkValue } from './precondition.js';
-import { cloneCircuitValue } from './circuit-value.js';
-import { JsonProof, verify } from './proof-system.js';
 import { invalidTransactionError } from './mina/errors.js';
-import { Types, TypesBigint } from '../bindings/mina-transaction/types.js';
+import { Types } from '../bindings/mina-transaction/types.js';
 import { Account } from './mina/account.js';
-import { TransactionCost, TransactionLimits } from './mina/constants.js';
 import { prettifyStacktrace } from './errors.js';
-import { Ml } from './ml/conversion.js';
-import {
-  transactionCommitments,
-  verifyAccountUpdateSignature,
-} from '../mina-signer/src/sign-zkapp-command.js';
 import { NetworkId } from '../mina-signer/src/types.js';
 import { currentTransaction } from './mina/transaction-context.js';
 import {
@@ -33,13 +15,15 @@ import {
   setActiveInstance,
   Mina,
   defaultNetworkConstants,
+  reportGetAccountError,
+  verifyTransactionLimits,
+  defaultNetworkState,
+  filterGroups,
   type FeePayerSpec,
   type DeprecatedFeePayerSpec,
   type ActionStates,
   type NetworkConstants,
 } from './mina/mina-instance.js';
-import { SimpleLedger } from './mina/transaction-logic/ledger.js';
-import { assert } from './gadgets/common.js';
 import { type EventActionFilterOptions } from './mina/graphql.js';
 import {
   type Transaction,
@@ -50,11 +34,12 @@ import {
   newTransaction,
   createIncludedOrRejectedTransaction,
 } from './mina/transaction.js';
+import { LocalBlockchain } from './mina/local-blockchain.js';
 
 export {
   BerkeleyQANet,
-  Network,
   LocalBlockchain,
+  Network,
   currentTransaction,
   Transaction,
   PendingTransaction,
@@ -99,369 +84,6 @@ const Transaction = {
     return newTransaction(transaction, activeInstance.proofsEnabled);
   },
 };
-
-function reportGetAccountError(publicKey: string, tokenId: string) {
-  if (tokenId === TokenId.toBase58(TokenId.default)) {
-    return `getAccount: Could not find account for public key ${publicKey}`;
-  } else {
-    return `getAccount: Could not find account for public key ${publicKey} with the tokenId ${tokenId}`;
-  }
-}
-
-/**
- * A mock Mina blockchain running locally and useful for testing.
- */
-function LocalBlockchain({
-  proofsEnabled = true,
-  enforceTransactionLimits = true,
-  networkId = 'testnet' as NetworkId,
-} = {}) {
-  const slotTime = 3 * 60 * 1000;
-  const startTime = Date.now();
-  const genesisTimestamp = UInt64.from(startTime);
-  const ledger = Ledger.create();
-  let networkState = defaultNetworkState();
-  let minaNetworkId: NetworkId = networkId;
-
-  function addAccount(publicKey: PublicKey, balance: string) {
-    ledger.addAccount(Ml.fromPublicKey(publicKey), balance);
-  }
-
-  let testAccounts: {
-    publicKey: PublicKey;
-    privateKey: PrivateKey;
-  }[] = [];
-
-  for (let i = 0; i < 10; ++i) {
-    let MINA = 10n ** 9n;
-    const largeValue = 1000n * MINA;
-    const k = PrivateKey.random();
-    const pk = k.toPublicKey();
-    addAccount(pk, largeValue.toString());
-    testAccounts.push({ privateKey: k, publicKey: pk });
-  }
-
-  const events: Record<string, any> = {};
-  const actions: Record<
-    string,
-    Record<string, { actions: string[][]; hash: string }[]>
-  > = {};
-
-  return {
-    getNetworkId: () => minaNetworkId,
-    proofsEnabled,
-    /**
-     * @deprecated use {@link Mina.getNetworkConstants}
-     */
-    accountCreationFee: () => defaultNetworkConstants.accountCreationFee,
-    getNetworkConstants() {
-      return {
-        ...defaultNetworkConstants,
-        genesisTimestamp,
-      };
-    },
-    currentSlot() {
-      return UInt32.from(
-        Math.ceil((new Date().valueOf() - startTime) / slotTime)
-      );
-    },
-    hasAccount(publicKey: PublicKey, tokenId: Field = TokenId.default) {
-      return !!ledger.getAccount(
-        Ml.fromPublicKey(publicKey),
-        Ml.constFromField(tokenId)
-      );
-    },
-    getAccount(
-      publicKey: PublicKey,
-      tokenId: Field = TokenId.default
-    ): Account {
-      let accountJson = ledger.getAccount(
-        Ml.fromPublicKey(publicKey),
-        Ml.constFromField(tokenId)
-      );
-      if (accountJson === undefined) {
-        throw new Error(
-          reportGetAccountError(publicKey.toBase58(), TokenId.toBase58(tokenId))
-        );
-      }
-      return Types.Account.fromJSON(accountJson);
-    },
-    getNetworkState() {
-      return networkState;
-    },
-    async sendTransaction(txn: Transaction): Promise<PendingTransaction> {
-      txn.sign();
-
-      let zkappCommandJson = ZkappCommand.toJSON(txn.transaction);
-      let commitments = transactionCommitments(
-        TypesBigint.ZkappCommand.fromJSON(zkappCommandJson),
-        minaNetworkId
-      );
-
-      if (enforceTransactionLimits) verifyTransactionLimits(txn.transaction);
-
-      // create an ad-hoc ledger to record changes to accounts within the transaction
-      let simpleLedger = SimpleLedger.create();
-
-      for (const update of txn.transaction.accountUpdates) {
-        let authIsProof = !!update.authorization.proof;
-        let kindIsProof = update.body.authorizationKind.isProved.toBoolean();
-        // checks and edge case where a proof is expected, but the developer forgot to invoke await tx.prove()
-        // this resulted in an assertion OCaml error, which didn't contain any useful information
-        if (kindIsProof && !authIsProof) {
-          throw Error(
-            `The actual authorization does not match the expected authorization kind. Did you forget to invoke \`await tx.prove();\`?`
-          );
-        }
-
-        let account = simpleLedger.load(update.body);
-
-        // the first time we encounter an account, use it from the persistent ledger
-        if (account === undefined) {
-          let accountJson = ledger.getAccount(
-            Ml.fromPublicKey(update.body.publicKey),
-            Ml.constFromField(update.body.tokenId)
-          );
-          if (accountJson !== undefined) {
-            let storedAccount = Account.fromJSON(accountJson);
-            simpleLedger.store(storedAccount);
-            account = storedAccount;
-          }
-        }
-
-        // TODO: verify account update even if the account doesn't exist yet, using a default initial account
-        if (account !== undefined) {
-          let publicInput = update.toPublicInput(txn.transaction);
-          await verifyAccountUpdate(
-            account,
-            update,
-            publicInput,
-            commitments,
-            this.proofsEnabled,
-            this.getNetworkId()
-          );
-          simpleLedger.apply(update);
-        }
-      }
-
-      let isSuccess = true;
-      const errors: string[] = [];
-      try {
-        ledger.applyJsonTransaction(
-          JSON.stringify(zkappCommandJson),
-          defaultNetworkConstants.accountCreationFee.toString(),
-          JSON.stringify(networkState)
-        );
-      } catch (err: any) {
-        isSuccess = false;
-        try {
-          const errorMessages = JSON.parse(err.message);
-          const formattedError = invalidTransactionError(
-            txn.transaction,
-            errorMessages,
-            {
-              accountCreationFee:
-                defaultNetworkConstants.accountCreationFee.toString(),
-            }
-          );
-          errors.push(formattedError);
-        } catch (parseError: any) {
-          const fallbackErrorMessage =
-            err.message || parseError.message || 'Unknown error occurred';
-          errors.push(fallbackErrorMessage);
-        }
-      }
-
-      // fetches all events from the transaction and stores them
-      // events are identified and associated with a publicKey and tokenId
-      txn.transaction.accountUpdates.forEach((p, i) => {
-        let pJson = zkappCommandJson.accountUpdates[i];
-        let addr = pJson.body.publicKey;
-        let tokenId = pJson.body.tokenId;
-        events[addr] ??= {};
-        if (p.body.events.data.length > 0) {
-          events[addr][tokenId] ??= [];
-          let updatedEvents = p.body.events.data.map((data) => {
-            return {
-              data,
-              transactionInfo: {
-                transactionHash: '',
-                transactionStatus: '',
-                transactionMemo: '',
-              },
-            };
-          });
-          events[addr][tokenId].push({
-            events: updatedEvents,
-            blockHeight: networkState.blockchainLength,
-            globalSlot: networkState.globalSlotSinceGenesis,
-            // The following fields are fetched from the Mina network. For now, we mock these values out
-            // since networkState does not contain these fields.
-            blockHash: '',
-            parentBlockHash: '',
-            chainStatus: '',
-          });
-        }
-
-        // actions/sequencing events
-
-        // most recent action state
-        let storedActions = actions[addr]?.[tokenId];
-        let latestActionState_ =
-          storedActions?.[storedActions.length - 1]?.hash;
-        // if there exists no hash, this means we initialize our latest hash with the empty state
-        let latestActionState =
-          latestActionState_ !== undefined
-            ? Field(latestActionState_)
-            : Actions.emptyActionState();
-
-        actions[addr] ??= {};
-        if (p.body.actions.data.length > 0) {
-          let newActionState = Actions.updateSequenceState(
-            latestActionState,
-            p.body.actions.hash
-          );
-          actions[addr][tokenId] ??= [];
-          actions[addr][tokenId].push({
-            actions: pJson.body.actions,
-            hash: newActionState.toString(),
-          });
-        }
-      });
-
-      const hash = Test.transactionHash.hashZkAppCommand(txn.toJSON());
-      const pendingTransaction: Omit<
-        PendingTransaction,
-        'wait' | 'waitOrThrowIfError'
-      > = {
-        isSuccess,
-        errors,
-        transaction: txn.transaction,
-        toJSON: txn.toJSON,
-        toPretty: txn.toPretty,
-        hash: (): string => {
-          return hash;
-        },
-      };
-
-      const wait = async (_options?: {
-        maxAttempts?: number;
-        interval?: number;
-      }) => {
-        return createIncludedOrRejectedTransaction(
-          pendingTransaction,
-          pendingTransaction.errors
-        );
-      };
-
-      const waitOrThrowIfError = async (_options?: {
-        maxAttempts?: number;
-        interval?: number;
-      }) => {
-        return createIncludedOrRejectedTransaction(
-          pendingTransaction,
-          pendingTransaction.errors
-        );
-      };
-
-      return {
-        ...pendingTransaction,
-        wait,
-        waitOrThrowIfError,
-      };
-    },
-    async transaction(sender: DeprecatedFeePayerSpec, f: () => void) {
-      // bad hack: run transaction just to see whether it creates proofs
-      // if it doesn't, this is the last chance to run SmartContract.runOutsideCircuit, which is supposed to run only once
-      // TODO: this has obvious holes if multiple zkapps are involved, but not relevant currently because we can't prove with multiple account updates
-      // and hopefully with upcoming work by Matt we can just run everything in the prover, and nowhere else
-      let tx = createTransaction(sender, f, 0, {
-        isFinalRunOutsideCircuit: false,
-        proofsEnabled: this.proofsEnabled,
-        fetchMode: 'test',
-      });
-      let hasProofs = tx.transaction.accountUpdates.some(
-        Authorization.hasLazyProof
-      );
-      return createTransaction(sender, f, 1, {
-        isFinalRunOutsideCircuit: !hasProofs,
-        proofsEnabled: this.proofsEnabled,
-      });
-    },
-    applyJsonTransaction(json: string) {
-      return ledger.applyJsonTransaction(
-        json,
-        defaultNetworkConstants.accountCreationFee.toString(),
-        JSON.stringify(networkState)
-      );
-    },
-    async fetchEvents(publicKey: PublicKey, tokenId: Field = TokenId.default) {
-      return events?.[publicKey.toBase58()]?.[TokenId.toBase58(tokenId)] ?? [];
-    },
-    async fetchActions(
-      publicKey: PublicKey,
-      actionStates?: ActionStates,
-      tokenId: Field = TokenId.default
-    ) {
-      return this.getActions(publicKey, actionStates, tokenId);
-    },
-    getActions(
-      publicKey: PublicKey,
-      actionStates?: ActionStates,
-      tokenId: Field = TokenId.default
-    ): { hash: string; actions: string[][] }[] {
-      let currentActions =
-        actions?.[publicKey.toBase58()]?.[TokenId.toBase58(tokenId)] ?? [];
-      let { fromActionState, endActionState } = actionStates ?? {};
-
-      let emptyState = Actions.emptyActionState();
-      if (endActionState?.equals(emptyState).toBoolean()) return [];
-
-      let start = fromActionState?.equals(emptyState).toBoolean()
-        ? undefined
-        : fromActionState?.toString();
-      let end = endActionState?.toString();
-
-      let startIndex = 0;
-      if (start) {
-        let i = currentActions.findIndex((e) => e.hash === start);
-        if (i === -1) throw Error(`getActions: fromActionState not found.`);
-        startIndex = i + 1;
-      }
-      let endIndex: number | undefined;
-      if (end) {
-        let i = currentActions.findIndex((e) => e.hash === end);
-        if (i === -1) throw Error(`getActions: endActionState not found.`);
-        endIndex = i + 1;
-      }
-      return currentActions.slice(startIndex, endIndex);
-    },
-    addAccount,
-    /**
-     * An array of 10 test accounts that have been pre-filled with
-     * 30000000000 units of currency.
-     */
-    testAccounts,
-    setGlobalSlot(slot: UInt32 | number) {
-      networkState.globalSlotSinceGenesis = UInt32.from(slot);
-    },
-    incrementGlobalSlot(increment: UInt32 | number) {
-      networkState.globalSlotSinceGenesis =
-        networkState.globalSlotSinceGenesis.add(increment);
-    },
-    setBlockchainLength(height: UInt32) {
-      networkState.blockchainLength = height;
-    },
-    setTotalCurrency(currency: UInt64) {
-      networkState.totalCurrency = currency;
-    },
-    setProofsEnabled(newProofsEnabled: boolean) {
-      this.proofsEnabled = newProofsEnabled;
-    },
-  };
-}
-// assert type compatibility without preventing LocalBlockchain to return additional properties / methods
-LocalBlockchain satisfies (...args: any) => Mina;
 
 /**
  * Represents the Mina blockchain running on a real network
@@ -1009,317 +631,6 @@ function dummyAccount(pubkey?: PublicKey): Account {
   let dummy = Types.Account.empty();
   if (pubkey) dummy.publicKey = pubkey;
   return dummy;
-}
-
-function defaultNetworkState(): NetworkValue {
-  let epochData: NetworkValue['stakingEpochData'] = {
-    ledger: { hash: Field(0), totalCurrency: UInt64.zero },
-    seed: Field(0),
-    startCheckpoint: Field(0),
-    lockCheckpoint: Field(0),
-    epochLength: UInt32.zero,
-  };
-  return {
-    snarkedLedgerHash: Field(0),
-    blockchainLength: UInt32.zero,
-    minWindowDensity: UInt32.zero,
-    totalCurrency: UInt64.zero,
-    globalSlotSinceGenesis: UInt32.zero,
-    stakingEpochData: epochData,
-    nextEpochData: cloneCircuitValue(epochData),
-  };
-}
-
-async function verifyAccountUpdate(
-  account: Account,
-  accountUpdate: AccountUpdate,
-  publicInput: ZkappPublicInput,
-  transactionCommitments: { commitment: bigint; fullCommitment: bigint },
-  proofsEnabled: boolean,
-  networkId: NetworkId
-): Promise<void> {
-  // check that that top-level updates have mayUseToken = No
-  // (equivalent check exists in the Mina node)
-  if (
-    accountUpdate.body.callDepth === 0 &&
-    !AccountUpdate.MayUseToken.isNo(accountUpdate).toBoolean()
-  ) {
-    throw Error(
-      'Top-level account update can not use or pass on token permissions. Make sure that\n' +
-        'accountUpdate.body.mayUseToken = AccountUpdate.MayUseToken.No;'
-    );
-  }
-
-  let perm = account.permissions;
-
-  // check if addMissingSignatures failed to include a signature
-  // due to a missing private key
-  if (accountUpdate.authorization === dummySignature()) {
-    let pk = PublicKey.toBase58(accountUpdate.body.publicKey);
-    throw Error(
-      `verifyAccountUpdate: Detected a missing signature for (${pk}), private key was missing.`
-    );
-  }
-  // we are essentially only checking if the update is empty or an actual update
-  function includesChange<T extends {}>(
-    val: T | string | null | (string | null)[]
-  ): boolean {
-    if (Array.isArray(val)) {
-      return !val.every((v) => v === null);
-    } else {
-      return val !== null;
-    }
-  }
-
-  function permissionForUpdate(key: string): Types.AuthRequired {
-    switch (key) {
-      case 'appState':
-        return perm.editState;
-      case 'delegate':
-        return perm.setDelegate;
-      case 'verificationKey':
-        return perm.setVerificationKey.auth;
-      case 'permissions':
-        return perm.setPermissions;
-      case 'zkappUri':
-        return perm.setZkappUri;
-      case 'tokenSymbol':
-        return perm.setTokenSymbol;
-      case 'timing':
-        return perm.setTiming;
-      case 'votingFor':
-        return perm.setVotingFor;
-      case 'actions':
-        return perm.editActionState;
-      case 'incrementNonce':
-        return perm.incrementNonce;
-      case 'send':
-        return perm.send;
-      case 'receive':
-        return perm.receive;
-      default:
-        throw Error(`Invalid permission for field ${key}: does not exist.`);
-    }
-  }
-
-  let accountUpdateJson = accountUpdate.toJSON();
-  const update = accountUpdateJson.body.update;
-
-  let errorTrace = '';
-
-  let isValidProof = false;
-  let isValidSignature = false;
-
-  // we don't check if proofs aren't enabled
-  if (!proofsEnabled) isValidProof = true;
-
-  if (accountUpdate.authorization.proof && proofsEnabled) {
-    try {
-      let publicInputFields = ZkappPublicInput.toFields(publicInput);
-
-      let proof: JsonProof = {
-        maxProofsVerified: 2,
-        proof: accountUpdate.authorization.proof!,
-        publicInput: publicInputFields.map((f) => f.toString()),
-        publicOutput: [],
-      };
-
-      let verificationKey = account.zkapp?.verificationKey?.data;
-      assert(
-        verificationKey !== undefined,
-        'Account does not have a verification key'
-      );
-
-      isValidProof = await verify(proof, verificationKey);
-      if (!isValidProof) {
-        throw Error(
-          `Invalid proof for account update\n${JSON.stringify(update)}`
-        );
-      }
-    } catch (error) {
-      errorTrace += '\n\n' + (error as Error).stack;
-      isValidProof = false;
-    }
-  }
-
-  if (accountUpdate.authorization.signature) {
-    // checking permissions and authorization for each account update individually
-    try {
-      isValidSignature = verifyAccountUpdateSignature(
-        TypesBigint.AccountUpdate.fromJSON(accountUpdateJson),
-        transactionCommitments,
-        networkId
-      );
-    } catch (error) {
-      errorTrace += '\n\n' + (error as Error).stack;
-      isValidSignature = false;
-    }
-  }
-
-  let verified = false;
-
-  function checkPermission(p0: Types.AuthRequired, field: string) {
-    let p = Types.AuthRequired.toJSON(p0);
-    if (p === 'None') return;
-
-    if (p === 'Impossible') {
-      throw Error(
-        `Transaction verification failed: Cannot update field '${field}' because permission for this field is '${p}'`
-      );
-    }
-
-    if (p === 'Signature' || p === 'Either') {
-      verified ||= isValidSignature;
-    }
-
-    if (p === 'Proof' || p === 'Either') {
-      verified ||= isValidProof;
-    }
-
-    if (!verified) {
-      throw Error(
-        `Transaction verification failed: Cannot update field '${field}' because permission for this field is '${p}', but the required authorization was not provided or is invalid.
-        ${errorTrace !== '' ? 'Error trace: ' + errorTrace : ''}\n\n`
-      );
-    }
-  }
-
-  // goes through the update field on a transaction
-  Object.entries(update).forEach(([key, value]) => {
-    if (includesChange(value)) {
-      let p = permissionForUpdate(key);
-      checkPermission(p, key);
-    }
-  });
-
-  // checks the sequence events (which result in an updated sequence state)
-  if (accountUpdate.body.actions.data.length > 0) {
-    let p = permissionForUpdate('actions');
-    checkPermission(p, 'actions');
-  }
-
-  if (accountUpdate.body.incrementNonce.toBoolean()) {
-    let p = permissionForUpdate('incrementNonce');
-    checkPermission(p, 'incrementNonce');
-  }
-
-  // this checks for an edge case where an account update can be approved using proofs but
-  // a) the proof is invalid (bad verification key)
-  // and b) there are no state changes initiate so no permissions will be checked
-  // however, if the verification key changes, the proof should still be invalid
-  if (errorTrace && !verified) {
-    throw Error(
-      `One or more proofs were invalid and no other form of authorization was provided.\n${errorTrace}`
-    );
-  }
-}
-
-function verifyTransactionLimits({ accountUpdates }: ZkappCommand) {
-  let eventElements = { events: 0, actions: 0 };
-
-  let authKinds = accountUpdates.map((update) => {
-    eventElements.events += countEventElements(update.body.events);
-    eventElements.actions += countEventElements(update.body.actions);
-    let { isSigned, isProved, verificationKeyHash } =
-      update.body.authorizationKind;
-    return {
-      isSigned: isSigned.toBoolean(),
-      isProved: isProved.toBoolean(),
-      verificationKeyHash: verificationKeyHash.toString(),
-    };
-  });
-  // insert entry for the fee payer
-  authKinds.unshift({
-    isSigned: true,
-    isProved: false,
-    verificationKeyHash: '',
-  });
-  let authTypes = filterGroups(authKinds);
-
-  /*
-  np := proof
-  n2 := signedPair
-  n1 := signedSingle
-  
-  formula used to calculate how expensive a zkapp transaction is
-
-  10.26*np + 10.08*n2 + 9.14*n1 < 69.45
-  */
-  let totalTimeRequired =
-    TransactionCost.PROOF_COST * authTypes.proof +
-    TransactionCost.SIGNED_PAIR_COST * authTypes.signedPair +
-    TransactionCost.SIGNED_SINGLE_COST * authTypes.signedSingle;
-
-  let isWithinCostLimit = totalTimeRequired < TransactionCost.COST_LIMIT;
-
-  let isWithinEventsLimit =
-    eventElements.events <= TransactionLimits.MAX_EVENT_ELEMENTS;
-  let isWithinActionsLimit =
-    eventElements.actions <= TransactionLimits.MAX_ACTION_ELEMENTS;
-
-  let error = '';
-
-  if (!isWithinCostLimit) {
-    // TODO: we should add a link to the docs explaining the reasoning behind it once we have such an explainer
-    error += `Error: The transaction is too expensive, try reducing the number of AccountUpdates that are attached to the transaction.
-Each transaction needs to be processed by the snark workers on the network.
-Certain layouts of AccountUpdates require more proving time than others, and therefore are too expensive.
-
-${JSON.stringify(authTypes)}
-\n\n`;
-  }
-
-  if (!isWithinEventsLimit) {
-    error += `Error: The account updates in your transaction are trying to emit too much event data. The maximum allowed number of field elements in events is ${TransactionLimits.MAX_EVENT_ELEMENTS}, but you tried to emit ${eventElements.events}.\n\n`;
-  }
-
-  if (!isWithinActionsLimit) {
-    error += `Error: The account updates in your transaction are trying to emit too much action data. The maximum allowed number of field elements in actions is ${TransactionLimits.MAX_ACTION_ELEMENTS}, but you tried to emit ${eventElements.actions}.\n\n`;
-  }
-
-  if (error) throw Error('Error during transaction sending:\n\n' + error);
-}
-
-function countEventElements({ data }: Events) {
-  return data.reduce((acc, ev) => acc + ev.length, 0);
-}
-
-type AuthorizationKind = { isProved: boolean; isSigned: boolean };
-
-const isPair = (a: AuthorizationKind, b: AuthorizationKind) =>
-  !a.isProved && !b.isProved;
-
-function filterPairs(xs: AuthorizationKind[]): {
-  xs: { isProved: boolean; isSigned: boolean }[];
-  pairs: number;
-} {
-  if (xs.length <= 1) return { xs, pairs: 0 };
-  if (isPair(xs[0], xs[1])) {
-    let rec = filterPairs(xs.slice(2));
-    return { xs: rec.xs, pairs: rec.pairs + 1 };
-  } else {
-    let rec = filterPairs(xs.slice(1));
-    return { xs: [xs[0]].concat(rec.xs), pairs: rec.pairs };
-  }
-}
-
-function filterGroups(xs: AuthorizationKind[]) {
-  let pairs = filterPairs(xs);
-  xs = pairs.xs;
-
-  let singleCount = 0;
-  let proofCount = 0;
-
-  xs.forEach((t) => {
-    if (t.isProved) proofCount++;
-    else singleCount++;
-  });
-
-  return {
-    signedPair: pairs.pairs,
-    signedSingle: singleCount,
-    proof: proofCount,
-  };
 }
 
 async function waitForFunding(address: string): Promise<void> {
